requirements.txt
pytorch-lightning==1.3.2
matplotlib==3.4.2
jupyter
git+https://github.com/firekind/athena